Crate ra_ide_db

This crate defines the core datastructure representing IDE state -- RootDatabase.

It is mainly a HirDatabase for semantic analysis, plus a SymbolsDatabase, for fuzzy search.

Modules

change
Defines a unit of change that can applied to a state of IDE to get the next state. Changes are transactional.

defs
NameDefinition keeps information about the element we want to search references for. The element is represented by NameKind. It's located inside some container and has a visibility, which defines a search scope. Note that the reference search is possible for not all of the classified items.

imports_locator
This module contains an import search funcionality that is provided to the ra_assists module. Later, this should be moved away to a separate crate that is accessible from the ra_assists module.

line_index
LineIndex maps flat TextSize offsets into (Line, Column) representation.

line_index_utils
Code actions can specify desirable final position of the cursor.

marks
See test_utils/src/marks.rs

search
Implementation of find-usages functionality.

symbol_index
This module handles fuzzy-searching of functions, structs and other symbols by name across the whole workspace and dependencies.

Structs

LineIndexDatabaseGroupStorage__

LineIndexDatabaseStorage
Representative struct for the query group.

LineIndexQuery

RootDatabase

Enums

LineIndexDatabaseGroupKey__

Traits

LineIndexDatabase