Module ra_ide_db::defs

NameDefinition keeps information about the element we want to search references for. The element is represented by NameKind. It's located inside some container and has a visibility, which defines a search scope. Note that the reference search is possible for not all of the classified items.

Enums

Definition

NameClass

NameRefClass

Functions

classify_name

classify_name_ref